implement RecordPosition service fn

+++ domain/geopoint.go
@@ -2,6 +2,6 @@ package domain
 
 // GeoPoint represents a geographical coordinate with latitude and longitude.
 type GeoPoint struct {
-	Latitude  float64 `json:"latitude"`
-	Longitude float64 `json:"longitude"`
+	Latitude  float32 `json:"latitude"`
+	Longitude float32 `json:"longitude"`
 }
